+++ src/cfg.rs
@@ -347,8 +347,11 @@ impl Host {
         }
     }
 
-    pub fn get_hostname(&self) -> &str {
-        self.hostname.as_ref().unwrap_or(&self.alias)
+    /// Get hostname as configured or a supplied default or (finally) alias name of the host.
+    pub fn get_hostname_def<'a>(&'a self, default: Option<String>) -> String {
+        self.hostname.clone()
+            .or(default)
+            .unwrap_or(self.alias.clone())
     }
 
     pub fn get_username(&self) -> String {