+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   This This allows for easier box copying of links when uploading several files at once.
 * `list`-command: When specifying `--detail`, fix files below one KiB having no unit suffix.
   Now the suffix is `B`.
+* If not specified explicitly, get hostname for host-entry from openSSH config.
 
 ## v0.7.1 (2021-01-26)